Remove duplicate and added endpoint description
